+++ src/main/java/net/frozenblock/wilderwild/registry/RegisterBlockEntities.java
@@ -45,7 +45,7 @@ public final class RegisterBlockEntities {
 
 	@NotNull
 	private static <T extends BlockEntity> BlockEntityType<T> register(@NotNull String path, BlockEntityType.@NotNull Builder<T> builder) {
-		Type<?> type = Util.fetchChoiceType(References.BLOCK_ENTITY, path);
+		Type<?> type = Util.fetchChoiceType(References.BLOCK_ENTITY, WilderConstants.string(path));
 		return Registry.register(BuiltInRegistries.BLOCK_ENTITY_TYPE, WilderConstants.id(path), builder.build(type));
 	}