+++ CHANGELOG.md
@@ -10,4 +10,5 @@ Put the changelog BELOW the dashes. ANYTHING ABOVE IS IGNORED
 - Converted Wilder Wild's entity registry to use Vanilla's entity builder.
 - Reorganized data generation for configured and placed features in caves into their own classes.
 - Changed the `wilderwild:blank_shut_up` feature into the `wilderwild:empty` feature, which now generates nothing instead of a single Water block.
+- Fixed Wilder Wild's block entities and entities causing log errors about the datafixer.
 - Fixed Mesoglea rendering with Sodium.